+++ src/components/SystemBar/OrigamSystemBar.vue
@@ -89,26 +89,26 @@
 		scoped
 >
 	.origam-system-bar {
-		align-items: center;
-		display: flex;
-		flex: 1 1 auto;
-		height: 24px;
-		justify-content: flex-end;
-		max-width: 100%;
-		padding-inline: 8px;
-		position: relative;
-		text-align: end;
-		width: 100%;
-		background: rgba(66, 66, 66, 1);
-		color: rgba(255, 255, 255, 1);
-		font-size: .75rem;
-		font-weight: 400;
-		letter-spacing: .0333333333em;
-		line-height: 1.667;
-		text-transform: none;
+		align-items: var(--origam-system-bar---align-items, center);
+		display: var(--origam-system-bar---display, flex);
+		flex: var(--origam-system-bar---flex, 1 1 auto);
+		height: var(--origam-system-bar---height, 24px);
+		justify-content: var(--origam-system-bar---justify-content, flex-end);
+		max-width: var(--origam-system-bar---max-width, 100%);
+		padding-inline: var(--origam-system-bar---padding-inline, 8px);
+		position: var(--origam-system-bar---position, relative);
+		text-align: var(--origam-system-bar---text-align, end);
+		width: var(--origam-system-bar---width, 100%);
+		background: var(--origam-system-bar---background, var(--origam-color-neutral-700, #404040));
+		color: var(--origam-system-bar---color, var(--origam-color-text-inverse, #FFFFFF));
+		font-size: var(--origam-system-bar---font-size, .75rem);
+		font-weight: var(--origam-system-bar---font-weight, 400);
+		letter-spacing: var(--origam-system-bar---letter-spacing, .0333333333em);
+		line-height: var(--origam-system-bar---line-height, 1.667);
+		text-transform: var(--origam-system-bar---text-transform, none);
 
 		.origam-icon {
-			opacity: 0.7;
+			opacity: var(--origam-system-bar__icon---opacity, 0.7);
 		}
 
 		&--absolute {
@@ -120,17 +120,12 @@
 		}
 
 		&--rounded {
-			border-radius: 0;
+			border-radius: var(--origam-system-bar--rounded---border-radius, 0);
 		}
 
 		&--window {
-			height: 32px;
+			height: var(--origam-system-bar---height-window, 32px);
 		}
 	}
 </style>
 
-<style>
-	:root {
-
-	}
-</style>